NPODObject

An NPOD object

Public fields

simulation

An ospsuite Simulation object.

optimizationParameterList

A list of parameters for optimization

outputPath

The path to the output of the simulation.

pkData

A data.frame of PK parameters

demographicData

A data.frame of demographic parameters

populationData

A data.frame of population parameters

npodRunSettings

A NPODRunSettings object

cached_mu

Numeric value of cached means

n_err

A data.frame of demographic parameters

theta_0

A data.frame of demographic parameters

pyl

A data.frame of demographic parameters

err_log

A data.frame of demographic parameters

sigma

A data.frame of demographic parameters

useLogNormalLikelihood

A logical to define if using log normal likelihood

Methods

Public methods

NPODObject$new()

NPODObject$parsePKData()

NPODObject$parsePopulationData()

NPODObject$setGrid()

NPODObject$runNPOD()

NPODObject$Dopt()

NPODObject$prune()

NPODObject$multi_D()

NPODObject$getPSI()

NPODObject$getLikelihood()

NPODObject$multi_mu()

NPODObject$clone()

Method new()

Initialize a new NPODObject

Usage

NPODObject$new(
  simulationFilePath,
  optimizationParameterList,
  outputPath,
  pkDataFilePath,
  studyPopulationDataFilePath = NULL,
  cofactorPaths = NULL,
  initialGridSize = NULL,
  useLogNormalLikelihood = NULL,
  npodRunSettings = NULL
)

Arguments

simulationFilePath

A string path to the simulation file

optimizationParameterList

A list of VBEParameter objects

outputPath

A string path to the output of the simulation

pkDataFilePath

A string path to the PK data file

studyPopulationDataFilePath

A string path to the study population data file

cofactorPaths

A list of cofactor paths

initialGridSize

An integer for the initial grid size

useLogNormalLikelihood

A logical to define if using log normal likelihood

npodRunSettings

A NPODRunSettings object

Method parsePKData()

Parse PK data and convert to base unit

Usage

NPODObject$parsePKData(pkDataFilePath)

Arguments

pkDataFilePath

A string path to the PK data file

Method parsePopulationData()

parse Population data and convert to base unit

Usage

NPODObject$parsePopulationData(
  studyPopulationDataFilePath,
  cofactorPaths = NULL
)

Arguments

studyPopulationDataFilePath

A string path to the study population data file

cofactorPaths

A list of cofactor paths

Method setGrid()

setGrid Set Grid for optimization

Usage

NPODObject$setGrid(initialGridSize)

Arguments

initialGridSize

An integer for the initial grid size

Method runNPOD()

Run NPOD

Usage

NPODObject$runNPOD()

Method Dopt()

Dopt Optimization algorithm

Usage

NPODObject$Dopt()

Method prune()

The prune function checks if a candidate support point (theta_plus) is sufficiently different from the existing set of support points (theta) and within specified bounds (a, b). If these conditions are met, theta_plus is added as a new column to theta.

Usage

NPODObject$prune(theta, theta_plus)

Arguments

theta

A matrix of support points

theta_plus

A candidate support point

Method multi_D()

multi_D computes a likelihood-based metric, D_comp, by summing the probabilities from getPSI(), normalized by self$pyl. It starts with a penalty term of -N, calls getPSI(), and updates D_comp.

Usage

NPODObject$multi_D(theta_parameter)

Arguments

theta_parameter

A matrix of support points

Method getPSI()

getPIS The getPSI function calculates the probability of the observed data mprob is an N x K matrix where each entry [i, l] represents the probability of observing y[[i]] given the model prediction ySimList[[i, l]]. A small lower bound, 1e-100, is enforced for numerical stability.

Usage

NPODObject$getPSI(
  theta,
  ySimList = NULL,
  useLog = self$useLogNormalLikelihood,
  optimizeSigma = FALSE,
  lambda = NULL
)

Arguments

theta

A matrix of support points

ySimList

A list of model predictions

useLog

A logical to define if using log normal likelihood

optimizeSigma

A logical to define if optimizing sigma

lambda

A list of lambda values

Method getLikelihood()

Calculates the likelihood of the observed data, y, given the model predictions, ySim, and measurement error, sigma. The default case uses arithmetic error, but if use_log = TRUE, it calculates the likelihood using the logarithmic (ratio) error.

Usage

NPODObject$getLikelihood(y, t, theta, sigma, ySim, useLog = FALSE)

Arguments

y

A vector of observed data

t

A vector of time points

theta

A matrix of support points

sigma

A vector of measurement errors

ySim

A vector of model predictions

useLog

A logical to define if using log normal likelihood

Method multi_mu()

computes the model predictions for each individual in the study population given a set of support points, theta.

Usage

NPODObject$multi_mu(theta)

Arguments

theta

A matrix of support points

Method clone()

The objects of this class are cloneable with this method.

Usage

NPODObject$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.