Calculation of BE statistics

Generates bioequivalence statistics

calc_be(
  data,
  n_trials = 50,
  subj_min = 5,
  subj_max = 50,
  subj_step = 5,
  seed = -17,
  ci_lcut = 0.8,
  ci_ucut = 1.25
)

Arguments

data

A dataframe or tibble with these columns: subject id, time, concentrations, period, drug.

n_trials

Integer coding the number of random trials per sample size.

subj_min

Integer defining the minimum number of subjects per trial

subj_max

Integer defining the maximum number of subjects per trial

subj_step

Integer defining the step size between subj_min and subj_max, so that the program will test sample sizes $N = {subj_min, subj_min + subj_step, subj_min + 2*subj_step,...,subj_max}$.

seed

Random number seed.

ci_lcut

Lower limit of the confidence interval.

ci_ucut

Upper limit of the confidence interval.

Value

A list of lists: [[1:n_samp]][[1:n_trials]][["par", "cross", "rep"]][[1,2]], where n_samp is the number of sample sizes tested, e.g., (subj_max - subj-min) / subj_step. The first item in the design is auc, and the second is cmax. For example the comparison of auc using the third trial of the second sample size for the replicated design would be x[[2]][[3]]$rep[[1]].

Details

Generates BE statistics for parallel, cross-over and full replication designs. From simulated data with 1000 subjects per drug, the function first calculates area under the concentration-time curve (AUC) and maximum concentration (Cmax) for each simulated profiles. Periods are defined as >1 for each replicated administration (e.g., when intra-occasion varibility is simulated.) Linear regression (stats::lm()) is used on log10(auc) andlog10(cmax) for parallel and cross-over designs, and mixed effect modeling (lmerTest::lmer()) is used for the replicated design.

Author

Michael Neely

Abdullah Hamadeh