Extract BE statistics

Extracts BE statistics from results of (calc_BE()).

extract_be(lmer_res, ci = 0.9, ci_lcut = 0.8, ci_ucut = 1.25)

Arguments

lmer_res

The output of (calc_BE()).

ci

Confidence interval for geometric mean ratio, default is 0.9 (90%) per FDA guidance.

ci_lcut

Lower confidence interval cutoff for geometric mean AUC or Cmax to be considered BE, default 0.8 per FDA guidance.

ci_ucut

Upper confidence interval cutoff for geometric mean AUC or Cmax to be considered BE, default 1.25 per FDA guidance.

Value

A list of lists: [[1:n_samp]][[1:n_trials]][["par", "cross", "rep"]][[1,2]], where n_samp is the number of sample sizes tested, e.g., (subj_max - subj-min) / subj_step. The first item in the design is auc, and the second is cmax. For example the comparison of auc using the third trial of the second sample size for the replicated design would be x[[2]][[3]]$rep[[1]].

Author

Michael Neely

Abdullah Hamadeh