Compute Weighted Support Points

Computes weighted support points using posterior probabilities and demographic parameters.

getWeightedPoints(res, params = c("Age", "Weight", "Height"))

Arguments

res

A list containing theta (a matrix of support points), w (weights), and demographicData (a data frame of demographic information).

params

A character vector of parameter names to extract from demographicData. Default is c('Age', 'Weight', 'Height').

Value

A list with points, a data frame of weighted support points, and weights, a vector of posterior probabilities.

Examples

res <- list(
theta = matrix(runif(9), 3, 3), 
w = c(0.2, 0.5, 0.3), 
demographicData = data.frame(Age = c(25, 30, 35), 
Weight = c(70, 80, 90), Height = c(170, 175, 180))
)
getWeightedPoints(res)
#> $points
#>       theta1      theta2    theta3 Age Weight Height
#> 1 0.08075014 0.834333037 0.6007609  25     70    170
#> 2 0.08075014 0.834333037 0.6007609  30     80    175
#> 3 0.08075014 0.834333037 0.6007609  35     90    180
#> 4 0.15720844 0.007399441 0.4663935  25     70    170
#> 5 0.15720844 0.007399441 0.4663935  30     80    175
#> 6 0.15720844 0.007399441 0.4663935  35     90    180
#> 7 0.49777739 0.289767245 0.7328820  25     70    170
#> 8 0.49777739 0.289767245 0.7328820  30     80    175
#> 9 0.49777739 0.289767245 0.7328820  35     90    180
#> 
#> $weights
#> [1] 0.06616492 0.64803446 0.20961155 0.32203301 0.01436804 0.40682353 0.61180207
#> [8] 0.33759749 0.38356492
#>